Change MOUNTVOL to SAVEDIR
Misleading ENV variable name.
It is supposed to hold the dir name where the zip files will be saved.

In fact, if one tries to mount a host volume to the MOUNTVOL,
the content of that dir will not be available.
Meaning that, one can not copy the zip back to the host.

So, the name changed as it is not meant to be mounted as a volume.